Adds post-commit hook to run hugo automatically
If you install it locally, it will run Hugo as long as you modify
anything in the `content` directory. If you use it within Travis,
something else will have to be done if it's not deployed to GitHub,
like adding API keys or say where it's going to be deployed.

Addresses issue #20. Maybe closes it, but I'm not sure.