Generates commit message with changed files
This addresses #20. It would have to be way more complicated if new
Hugo files had to be checked out. Maybe it's not a big deal, but if
that's the case it will print a message so that they can be added
manually.

I'm kind of wary of adding a push here... You neendn't push after each
commit, and it's probably reasonable to just leave the commits there
for editing if needed.

Also: commits could be squashed... Left as future work and if
absolutely necessary.