Use default templates for allcontributors
As stated in https://github.com/codeisscience/code-is-science/issues/87#issuecomment-686588245, defining a stub `contributorTemplate` have side effects such as taking off the emojis (since they're not part of the template). Since we're not using anything specific in the template (instead, we're just using a standard minimalist example template), we can rely on the default allcontributors' template parameters to generate the default HTML with emojis back.